refactor: split into dedicated circle component
